// src/generated/client/types/PostRoutinesRequestBody.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { PostRoutinesRequestExercise } from "./PostRoutinesRequestExercise.ts";

export type PostRoutinesRequestBody = {
    /**
     * @type object | undefined
    */
    routine?: {
        /**
         * @description The title of the routine.
         * @type string | undefined
        */
        title?: string;
        /**
         * @description The folder id the routine should be added to. Pass null to insert the routine into default \"My Routines\" folder
         * @type number
        */
        folder_id?: number | null;
        /**
         * @description Additional notes for the routine.
         * @type string | undefined
        */
        notes?: string;
        /**
         * @type array | undefined
        */
        exercises?: PostRoutinesRequestExercise[];
    };
};