// src/generated/client/types/PutRoutinesRequestBody.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { PutRoutinesRequestExercise } from "./PutRoutinesRequestExercise.ts";

export type PutRoutinesRequestBody = {
  /**
   * @type object | undefined
   */
  routine?: {
    /**
     * @description The title of the routine.
     * @type string | undefined
     */
    title?: string;
    /**
     * @description Additional notes for the routine.
     * @type string
     */
    notes?: string | null;
    /**
     * @type array | undefined
     */
    exercises?: PutRoutinesRequestExercise[];
  };
};